modify order script
