+++ src/test/resources/dataPopulation/datiNotifica.yaml
@@ -1,4 +1,4 @@
-numeroProtocollo: "TA-FFSMRC-20230719-6"
+numeroProtocollo: "TA-FFSMRC-20230807-3"
 oggettoDellaNotifica: "Pagamento rata IMU"
 descrizione: "PAGAMENTO RATA IMU"
 gruppoTest: "prova3"
