+++ src/test/resources/feature/2-mittente/5_apiKey/3_022_generaApiKeySenzaGruppo.feature
