+++ src/test/resources/feature/3-destinatario/personaFisica/deleghePF/3_046_visualizzaCodiceDelega.feature
@@ -13,7 +13,7 @@ Feature: Il persona fisica visualizza il codice  di una delega
     When Nella pagina Piattaforma Notifiche persona fisica click sul bottone Deleghe
     And Nella pagina Piattaforma Notifiche persona fisica si vede la sezione Deleghe
     And Nella sezione Deleghe si verifica sia presente una delega
-    And Nella sezione Deleghe si clicca sul menu della delega
+    And Nella sezione Deleghe si clicca sul menu della delega "nuova_delega"
     And Nella sezione Deleghe si sceglie l'opzione mostra codice
     Then Si clicca sul bottone chiudi
     And Logout da portale persona fisica
